requirements.txt
django
djangorestframework
requests
psycopg2
psycopg2-binary
djangorestframework-simplejwt